requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.17
jinja2==3.1.4
httpx==0.27.2
python-dotenv==1.0.1
pypdf==5.1.0
cryptography>=3.1
pdfplumber==0.11.4
pymupdf==1.24.10
pdf2image==1.17.0
pytesseract==0.3.10
Pillow>=10.0.0
